refactor: 도메인별로 폴더구성하여 응집도 높이기
