+++ src/app/api/scrap/constants/error/error-messages.ts
