+++ src/app/api/scrap/route.ts
@@ -1,4 +1,4 @@
-import { getAllScrapList } from "@/app/home/apis/scrap-apis";
+import { getAllScrapList } from "@/app/api/scrap/apis";
 import { NextResponse } from "next/server";
 
 export async function GET() {
