+++ src/app/blog/[slug]/page.tsx
@@ -1,5 +1,5 @@
 import { getAllCategoriesID, getFilteredPosts } from "@/lib/api";
-import TagNavbar from "@/app/home/components/tag-navbar/tag-navbar";
+import TagNavbar from "@/app/blog/components/tag-navbar/tag-navbar";
 import PostCards from "@/app/shared/_components/post/post-cards/post-cards";
 
 type Params = Promise<{ slug: string }>;
