+++ src/app/shared/_components/post/interfaces/posts.ts
